move TouchableParticle outside of RegionParticleIteratorTest
